Move transaction methods alongside their corresponding read logic.
